+++ src/main/kotlin/net/pwall/json/kotlin/codegen/gradle/input/InputComposite.kt
@@ -1,8 +1,8 @@
 /*
- * @(#) SchemaExtensionFormatValidation.kt
+ * @(#) InputComposite.kt
  *
  * json-kotlin-gradle  Gradle Code Generation Plugin for JSON Schema
- * Copyright (c) 2021 Peter Wall
+ * Copyright (c) 2022 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,35 +23,40 @@
  * SOFTWARE.
  */
 
-package net.pwall.json.kotlin.codegen.gradle.extension
+package net.pwall.json.kotlin.codegen.gradle.input
 
-import java.net.URI
+import java.io.File
 import javax.inject.Inject
 
 import org.gradle.api.Project
 import org.gradle.api.tasks.Input
+import org.gradle.kotlin.dsl.listProperty
 import org.gradle.kotlin.dsl.property
 
-import net.pwall.json.pointer.JSONPointer
-import net.pwall.json.schema.JSONSchema
-import net.pwall.json.schema.validation.FormatValidator
+import net.pwall.json.kotlin.codegen.gradle.JSONSchemaCodegenTask.Companion.addPointerTargets
+import net.pwall.json.schema.codegen.CodeGenerator
 
-class SchemaExtensionFormatValidation @Inject constructor(name: String, project: Project) :
-        SchemaExtension(name, project) {
+class InputComposite @Inject constructor(name: String, project: Project) : InputDefinition(name, project) {
 
     @Input
-    val format = project.objects.property<String>()
+    val file = project.objects.property<File>()
 
-    override fun validator(uri: URI?, pointer: JSONPointer): JSONSchema.Validator {
-        return FormatValidator(uri, pointer, when (format.get()) {
-            "date" -> FormatValidator.DateFormatChecker
-            "date-time" -> FormatValidator.DateTimeFormatChecker
-            "time" -> FormatValidator.TimeFormatChecker
-            "uri" -> FormatValidator.URIFormatChecker
-            "uri-reference" -> FormatValidator.URIReferenceFormatChecker
-            "uuid" -> FormatValidator.UUIDFormatChecker
-            else -> throw IllegalArgumentException("Invalid format type")
-        })
+    @Input
+    val pointer = project.objects.property<String>()
+
+    @Input
+    val include = project.objects.listProperty<String>()
+
+    @Input
+    val exclude = project.objects.listProperty<String>()
+
+    override fun applyTo(codeGenerator: CodeGenerator) {
+        file.orNull?.let {
+            val includes = include.orNull ?: emptyList()
+            val excludes = exclude.orNull ?: emptyList()
+            val ptr = pointer.orNull ?: throw IllegalArgumentException("No pointer specified")
+            codeGenerator.addPointerTargets(it, ptr, includes, excludes)
+        } ?: throw IllegalArgumentException("No File specified")
     }
 
 }